Fixed inventory issue
Overall issue was the forgetting of 'Tab' and 'E' being open inventory and interact with chest. The recent changes to the code lead me to belive that pressing 'E' would open the inventory, however in reality, 'E' is just the interact button to open the chest. The issue was complicated when attempts were made to integrate the rigged and animated character. This process resulted in what I belive to be a mix up, leading to the test_inv.tres file being removed from the player node. Suspicions initially arose as, when forcing the game to continue in debug mode only resulted in the icons in the player inventory missing. Leading me to belive that the supposed 'inventory_interact' error was a bogus error and was masking the real issue, which turned out to simply be a file not added to a node. Next steps are to complete the pickup item section of the video and integrate the rigged and animated solider.