Added pickaxe to models hand
But it doesn't appear in game. Need to fix audio as walking SFX try's to play each time a step is taken, resulting in the buzzing noise as they all try to play on top of each other. Considering creating a separate immovable rockfall animation for locks in's. Almost finished with corridors, might add in a few more, and then fix up floors. Considering camera animations for pick up's and arriving at the bunker door.