Trying to fix audio
Spend much longer than normal, due to starting day with work, rather than starting it halfway through. I managed to make good progress on the audio system, before it broke again. Used .ogg files instead of .wav as it is easier to loop. Issue is that now the buzzing noise is playing instead of the actual sfx, this could be due to Godot rapidly trying to play the sfx over and over. Will investigate and hopefully come back tonight and ultimately put this issue to bed.