Successfully implemented rockfall block in feature!
After around 4 hrs of work, I've done it!

The mechanic works on the rail principal of a Tripcock, if the player passes over the 'Tripcock' (collision object), it triggers an animation to play. This animation shows the rocks falling and also moves collision shapes from their elevated position to down onto the ground, similar to a castle gate moving vertically down. The tripcock also changes a global var so that the player has no choice but to continue as the rockfall won't clear due to them have insufficient 'permissions'.

The next steps are to implement this where the player moves into a new 'sector'.